Move GraphSidebar to atoms
