+++ frontend/src/components/atoms/GraphSidebar.tsx
