+++ frontend/src/components/molecules/GraphContainer.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Flex } from '@chakra-ui/react';
 import Graph from '../atoms/Graph';
-import GraphSidebar from './GraphSidebar';
+import GraphSidebar from '../atoms/GraphSidebar';
 
 const GraphContainer = () => (
   <Flex>
